% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_symbols.R
\name{get_symbols}
\alias{get_symbols}
\title{Download Stock Data From Yahoo}
\usage{
get_symbols(sym, ...)
}
\arguments{
\item{sym}{character; Symbols to be retrieved.}

\item{...}{Additional parameters passed to \link[tidyquant]{tq_get}}
}
\description{
A wrapper function for \link[tidyquant]{tq_get} to retrieve symbols from
Yahoo! Finance as \link[tsibble]{tsibble} time series.
}
